requirements.txt
apache-beam==2.49.0
apache-beam[gcp]==2.49.0
apache-airflow[gcp]==2.1.4
requests-toolbelt==1.0.0
google-cloud-pubsub==2.17.0

functions-framework==3.2.0
requests==2.26.0

google-auth==1.35.0


